ci: :bug: Nylon, Metal & vislon bug fixes
