+++ src/pages/Metal/Finishing/Log/index.jsx
@@ -1,9 +1,6 @@
 import { useEffect } from 'react';
 import RMTransferLog from './RMTransferLog';
-import SFGTransferLog from './SFGTransferLog';
-import RMOrderAgainstLog from './RMOrderAgainstLog';
-import ProductionLog from './ProductionLog';
-import TransactionLog from './TransactionLog';
+import Production from './Production';
 
 export default function Index() {
 	useEffect(() => {
@@ -11,13 +8,9 @@ export default function Index() {
 	}, []);
 	return (
 		<div>
-			<ProductionLog />
-			<hr className='my-6 border-2 border-dashed border-secondary/30' />
-			{/* <TransactionLog /> */}
+			<Production />
 			<hr className='my-6 border-2 border-dashed border-secondary/30' />
 			<RMTransferLog />
-			{/* <hr className='my-6 border-2 border-dashed border-secondary-content' />
-			<RMOrderAgainstLog /> */}
 		</div>
 	);
 }
