+++ src/pages/Nylon/PlasticFinishing/Production/index.jsx
@@ -74,7 +74,7 @@ export default function Index() {
 				cell: (info) => Number(info.getValue()),
 			},
 			{
-				accessorKey: 'nylon_plastic_finishing',
+				accessorKey: 'tape_transferred',
 				header: 'Tape Stock (KG)',
 				enableColumnFilter: false,
 				cell: (info) => Number(info.getValue()),
